codigo ordenado
