+++ src/index.html
@@ -21,25 +21,26 @@
         <!-- menu -->
     <nav class="nav" id="nav">
           <ul class="menu">
-            <li class="menu_item  menu_item--hover">
+            <li class="menu__item  menu__item--hover">
               <a href="" class="menu__link ">Filtrar</a>
               <ul class="submenu">
-                <li class="menu_item" id="fighter"><a href="" class="menu_link">Fighter</a></li>
-                <li class="menu_item" id="tank"><a href="" class="menu_link">Tank</a></li>
-                <li class="menu_item" id="mage"><a href="" class="menu_link">Mage</a></li>
-                <li class="menu_item" id="assasin"><a href="" class="menu_link">Assassin</a></li>
-                <li class="menu_item" id="support"><a href="" class="menu_link">Support</a></li>
-                <li class="menu_item" id="marksman"><a href="" class="menu_link">Marksman</a></li>
+                <li class="menu__item" id="fighter"><a href="#" class="menu__link">Fighter</a></li>
+                <li class="menu__item" id="tank"><a href="#" class="menu__link">Tank</a></li>
+                <li class="menu__item" id="mage"><a href="#" class="menu__link">Mage</a></li>
+                <li class="menu__item" id="assasin"><a href="#" class="menu__link">Assassin</a></li>
+                <li class="menu__item" id="support"><a href="#" class="menu__link">Support</a></li>
+                <li class="menu__item" id="marksman"><a href="#" class="menu__link">Marksman</a></li>
               </ul>
             </li>
-            <li class="menu_item menu_item--hover">
+            <li class="menu__item  menu__item--hover">
               <a href="" class="menu__link ">Ordenar</a>
               <ul class="submenu">
-                <li class="menu_item"><a href="" class="menu_link">De la A a la Z</a></li>
-                <li id="za" class="menu_item"><a href="" class="menu_link">De la Z a la A</a></li>
+                <li class="menu__item" id="az"><a href="#" class="menu__link">A-Z</a></li>
+                <li class="menu__item" id="za"><a href="#" class="menu__link">Z-A</a></li>
+
               </ul>
             </li>
-            <li class="menu_item"><a href="" class="menu_link">Calcular</a></li>
+            <li class="menu__item"><a href="#" class="menu__link">Calcular</a></li>
           </ul>
         </nav>
   </nav>
@@ -64,15 +65,15 @@
           <div class="gallery__card">
             <div class="gallery__info">
               <img src="" alt="" class="gallery__img"/>
-                <h2 class="gallery__name"></h2>
-                <h4 class="gallery__tags"></h4>
+                <h2 class="gallery__name">Katte</h2>
+                <h4 class="gallery__tags">Roles</h4>
             </div>
             <div class="gallery__reverse">
-              <h3 class="gallery__surname"></h3><br>
-              <h4 class="gallery__attack"></h4>
-              <h4 class="gallery__defense"></h4>
-              <h4 class="gallery__magic"></h4>
-              <h4 class="gallery__difficulty"></h4>
+              <h3 class="gallery__surname">Fer</h3><br>
+              <h4 class="gallery__attack">Ataque</h4>
+              <h4 class="gallery__defense">Defensa</h4>
+              <h4 class="gallery__magic">Magia</h4>
+              <h4 class="gallery__difficulty">Dificultad</h4>
             </div>
           </div>
         </template>
