+++ src/script/data.js
@@ -1,0 +1,28 @@
+
+export function peticion(data){
+
+  const championUser = () => {
+      const contenedor = document.getElementById("gallery")
+    const fragment = document.createDocumentFragment()
+    const templateGallery = document.getElementById('template-gallery').content;
+
+    const championRandom = data.slice().sort(() => Math.random() - 0.5);
+
+    for (const item of championRandom) {
+      templateGallery.querySelector('.gallery__img').src = item.splash
+      templateGallery.querySelector('.gallery__name').textContent=item.name
+      templateGallery.querySelector('.gallery__surname').textContent=item.title.toUpperCase()
+      templateGallery.querySelector('.gallery__attack').textContent=('Attack: ' + (item.info.attack))
+      templateGallery.querySelector('.gallery__defense').textContent=('Defense: ' + (item.info.defense))
+      templateGallery.querySelector('.gallery__magic').textContent=('Magic: ' + (item.info.magic))
+      templateGallery.querySelector('.gallery__difficulty').textContent=('Difficulty: ' + (item.info.difficulty))
+      templateGallery.querySelector('.gallery__tags').textContent=item.tags
+
+      const clone = templateGallery.cloneNode(true)
+      fragment.append(clone)
+    }
+    contenedor.append(fragment)
+    }
+    championUser()
+}
+
