+++ src/script/filtros.js
@@ -1,0 +1,100 @@
+    export const filter = () => {
+    const fighterBtn = document.getElementById('fighter');
+    const tankBtn = document.getElementById('tank');
+    const mageBtn = document.getElementById('mage');
+    const assasinBtn = document.getElementById('assasin');
+    const supportBtn = document.getElementById('support');
+    const marksmanBtn = document.getElementById('marksman');
+// const zaBtn = document.getElementById("az")
+
+// zaBtn.addEventListener('click',() => {
+//   let cards = document.querySelectorAll(".gallery__card")
+//       for (const card of cards) {
+//         let champion = card.textContent
+//         if (!champion.includes("Fighter")){
+//           card.classList.add("ocultar")
+//         }else{
+//           card.classList.remove("ocultar")
+//         }
+//       }
+// })
+
+
+    fighterBtn.addEventListener("click",() => {
+      let cards = document.querySelectorAll(".gallery__card")
+      for (const card of cards) {
+        let champion = card.textContent
+        if (!champion.includes("Fighter")){
+          card.classList.add("ocultar")
+        }else{
+          card.classList.remove("ocultar")
+        }
+      }
+    })
+
+    tankBtn.addEventListener("click",() => {
+      let cards = document.querySelectorAll(".gallery__card")
+      for (const card of cards) {
+        let champion = card.textContent
+
+        if (!champion.includes("Tank")){
+          card.classList.add("ocultar")
+        }else{
+          card.classList.remove("ocultar")
+        }
+      }
+    })
+
+    mageBtn.addEventListener("click",() => {
+      let cards = document.querySelectorAll(".gallery__card")
+      for (const card of cards) {
+        let champion = card.textContent
+
+        if (!champion.includes("Mage")){
+          card.classList.add("ocultar")
+        }else{
+          card.classList.remove("ocultar")
+        }
+      }
+    })
+
+    assasinBtn.addEventListener("click",() => {
+      let cards = document.querySelectorAll(".gallery__card")
+      for (const card of cards) {
+        let champion = card.textContent
+
+        if (!champion.includes("Assassin")){
+          card.classList.add("ocultar")
+        }else{
+          card.classList.remove("ocultar")
+        }
+      }
+    })
+
+    supportBtn.addEventListener("click",() => {
+      let cards = document.querySelectorAll(".gallery__card")
+      for (const card of cards) {
+        let champion = card.textContent
+
+        if (!champion.includes("Support")){
+          card.classList.add("ocultar")
+        }else{
+          card.classList.remove("ocultar")
+        }
+      }
+    })
+
+    marksmanBtn.addEventListener("click",() => {
+      let cards = document.querySelectorAll(".gallery__card")
+      for (const card of cards) {
+        let champion = card.textContent
+
+        if (!champion.includes("Marksman")){
+          card.classList.add("ocultar")
+        }else{
+          card.classList.remove("ocultar")
+        }
+      }
+    })
+
+}
